Improved installation script

+++ setup-cpputest.sh
@@ -1,16 +1,29 @@
 #!/bin/bash
 
-# Download CppuTest from Github
-rm -rf cpputest
-rm cpputest-4.0.tar.gz
-wget https://github.com/cpputest/cpputest/releases/download/v4.0/cpputest-4.0.tar.gz
-tar -xf cpputest-4.0.tar.gz
-rm cpputest-4.0.tar.gz
-mv cpputest-4.0 cpputest
-cd cpputest
+if [[ $1 = "-uninstall" ]] || [[ $1 = "-reinstall" ]]; then
+	rm -rf cpputest
+	rm cpputest-4.0.tar.gz
+fi
 
-# Actually install CppUTest
-# Needs automake and autoconf !
-./autogen.sh
-./configure
-make all
+
+if [[ $1 != "-uninstall" ]]; then
+
+	# Download CppuTest from Github
+	if [ ! -d "cpputest" ]; then
+		wget https://github.com/cpputest/cpputest/releases/download/v4.0/cpputest-4.0.tar.gz
+		tar -xf cpputest-4.0.tar.gz
+		rm cpputest-4.0.tar.gz
+		mv cpputest-4.0 cpputest
+	else
+		echo "cpputest already downloaded!"
+	fi
+
+	cd cpputest
+
+	# Actually install CppUTest
+	# Needs automake and autoconf !
+	./autogen.sh
+	./configure
+	make all
+
+fi
